Simplify the JAXB classes, by combining the 06 and 07 classes into one
This duplication was causing a bug within the COPCMessageHandler where
the handler was trying to unmarshall an 07 message into an 06 class.

By combining the two classes into one we remove this whole class of
defect.
