+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/service/NackAckPreparationService.java
@@ -4,7 +4,6 @@ import static uk.nhs.adaptors.common.enums.MigrationStatus.COPC_ACKNOWLEDGED;
 import static uk.nhs.adaptors.common.enums.MigrationStatus.FINAL_ACK_SENT;
 
 import org.hl7.v3.COPCIN000001UK01Message;
-import org.hl7.v3.RCMRIN030000UK06Message;
 import org.hl7.v3.RCMRIN030000UKMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,7 +27,7 @@ public class NackAckPreparationService implements NackAckPrepInterface {
     private final SendACKMessageHandler sendACKMessageHandler;
     private final MigrationStatusLogService migrationStatusLogService;
 
-    public boolean sendAckMessage(RCMRIN030000UK06Message payload, String conversationId) {
+    public boolean sendAckMessage(RCMRIN030000UKMessage payload, String conversationId) {
 
         LOGGER.debug("Sending Final ACK message for Conversation ID: [{}]", conversationId);
 
@@ -55,8 +54,8 @@ public class NackAckPreparationService implements NackAckPrepInterface {
         ));
     }
 
-    private ACKMessageData prepareAckMessageData(RCMRIN030000UK06Message payload,
-                                                String conversationId) {
+    private ACKMessageData prepareAckMessageData(RCMRIN030000UKMessage payload,
+                                                 String conversationId) {
 
         String toOdsCode = XmlParseUtilService.parseToOdsCode(payload);
         String messageRef = XmlParseUtilService.parseMessageRef(payload);
