+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/service/FailedProcessHandlingServiceTest.java
@@ -16,7 +16,7 @@ import java.util.UUID;
 
 import org.hl7.v3.COPCIN000001UK01Message;
 import org.hl7.v3.II;
-import org.hl7.v3.RCMRIN030000UK06Message;
+import org.hl7.v3.RCMRIN030000UKMessage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -42,7 +42,7 @@ public class FailedProcessHandlingServiceTest {
     @Mock
     private SendNACKMessageHandler sendNACKMessageHandler;
     @Mock
-    private RCMRIN030000UK06Message ehrExtractMessage;
+    private RCMRIN030000UKMessage ehrExtractMessage;
     @Mock
     private COPCIN000001UK01Message copcMessage;
     @Mock
