--- schema/src/main/java/org/hl7/v3/RCMRIN030000UK06ControlActEvent.java
@@ -1,355 +1,0 @@
-
-package org.hl7.v3;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * &lt;p&gt;Java class for RCMR_IN030000UK06.ControlActEvent complex type.
- * 
- * &lt;p&gt;The following schema fragment specifies the expected content contained within this class.
- * 
- * &lt;pre&gt;
- * &amp;lt;complexType name="RCMR_IN030000UK06.ControlActEvent"&amp;gt;
- *   &amp;lt;complexContent&amp;gt;
- *     &amp;lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&amp;gt;
- *       &amp;lt;sequence&amp;gt;
- *         &amp;lt;element name="author" type="{urn:hl7-org:v3}MCAI_MT040101UK03.Author" minOccurs="0"/&amp;gt;
- *         &amp;lt;element name="author1" type="{urn:hl7-org:v3}MCAI_MT040101UK03.Author2" maxOccurs="2"/&amp;gt;
- *         &amp;lt;element name="reason" type="{urn:hl7-org:v3}MCAI_MT040101UK03.Reason" maxOccurs="unbounded" minOccurs="0"/&amp;gt;
- *         &amp;lt;element name="subject" type="{urn:hl7-org:v3}RCMR_IN030000UK06.Subject"/&amp;gt;
- *       &amp;lt;/sequence&amp;gt;
- *       &amp;lt;attribute name="type" type="{urn:hl7-org:v3}Classes" default="ControlAct" /&amp;gt;
- *       &amp;lt;attribute name="classCode" type="{urn:hl7-org:v3}ActClass" default="CACT" /&amp;gt;
- *       &amp;lt;attribute name="moodCode" type="{urn:hl7-org:v3}ActMood" default="EVN" /&amp;gt;
- *       &amp;lt;attribute name="typeID"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}oid" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="realmCode"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}cs" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="nullFlavor" type="{urn:hl7-org:v3}cs" /&amp;gt;
- *     &amp;lt;/restriction&amp;gt;
- *   &amp;lt;/complexContent&amp;gt;
- * &amp;lt;/complexType&amp;gt;
- * &lt;/pre&gt;
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "RCMR_IN030000UK06.ControlActEvent", propOrder = {
-    "author",
-    "author1",
-    "reason",
-    "subject"
-})
-public class RCMRIN030000UK06ControlActEvent {
-
-    protected MCAIMT040101UK03Author author;
-    @XmlElement(required = true)
-    protected List<MCAIMT040101UK03Author2> author1;
-    protected List<MCAIMT040101UK03Reason> reason;
-    @XmlElement(required = true)
-    protected RCMRIN030000UK06Subject subject;
-    @XmlAttribute(name = "type")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String type;
-    @XmlAttribute(name = "classCode")
-    protected List<String> classCode;
-    @XmlAttribute(name = "moodCode")
-    protected List<String> moodCode;
-    @XmlAttribute(name = "typeID")
-    protected List<String> typeID;
-    @XmlAttribute(name = "realmCode")
-    protected List<String> realmCode;
-    @XmlAttribute(name = "nullFlavor")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String nullFlavor;
-
-    /**
-     * Gets the value of the author property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link MCAIMT040101UK03Author }
-     *     
-     */
-    public MCAIMT040101UK03Author getAuthor() {
-        return author;
-    }
-
-    /**
-     * Sets the value of the author property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link MCAIMT040101UK03Author }
-     *     
-     */
-    public void setAuthor(MCAIMT040101UK03Author value) {
-        this.author = value;
-    }
-
-    /**
-     * Gets the value of the author1 property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the author1 property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getAuthor1().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link MCAIMT040101UK03Author2 }
-     * 
-     * 
-     */
-    public List<MCAIMT040101UK03Author2> getAuthor1() {
-        if (author1 == null) {
-            author1 = new ArrayList<>();
-        }
-        return this.author1;
-    }
-
-    /**
-     * Gets the value of the reason property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the reason property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getReason().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link MCAIMT040101UK03Reason }
-     * 
-     * 
-     */
-    public List<MCAIMT040101UK03Reason> getReason() {
-        if (reason == null) {
-            reason = new ArrayList<>();
-        }
-        return this.reason;
-    }
-
-    /**
-     * Gets the value of the subject property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link RCMRIN030000UK06Subject }
-     *     
-     */
-    public RCMRIN030000UK06Subject getSubject() {
-        return subject;
-    }
-
-    /**
-     * Sets the value of the subject property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link RCMRIN030000UK06Subject }
-     *     
-     */
-    public void setSubject(RCMRIN030000UK06Subject value) {
-        this.subject = value;
-    }
-
-    /**
-     * Gets the value of the type property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getType() {
-        if (type == null) {
-            return "ControlAct";
-        } else {
-            return type;
-        }
-    }
-
-    /**
-     * Sets the value of the type property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setType(String value) {
-        this.type = value;
-    }
-
-    /**
-     * Gets the value of the classCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the classCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getClassCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getClassCode() {
-        if (classCode == null) {
-            classCode = new ArrayList<>();
-        }
-        return this.classCode;
-    }
-
-    /**
-     * Gets the value of the moodCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the moodCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getMoodCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getMoodCode() {
-        if (moodCode == null) {
-            moodCode = new ArrayList<>();
-        }
-        return this.moodCode;
-    }
-
-    /**
-     * Gets the value of the typeID property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the typeID property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getTypeID().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getTypeID() {
-        if (typeID == null) {
-            typeID = new ArrayList<>();
-        }
-        return this.typeID;
-    }
-
-    /**
-     * Gets the value of the realmCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the realmCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getRealmCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getRealmCode() {
-        if (realmCode == null) {
-            realmCode = new ArrayList<>();
-        }
-        return this.realmCode;
-    }
-
-    /**
-     * Gets the value of the nullFlavor property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNullFlavor() {
-        return nullFlavor;
-    }
-
-    /**
-     * Sets the value of the nullFlavor property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNullFlavor(String value) {
-        this.nullFlavor = value;
-    }
-
-}
