--- schema/src/main/java/org/hl7/v3/RCMRIN030000UK07Message.java
@@ -1,456 +1,0 @@
-
-package org.hl7.v3;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * &lt;p&gt;Java class for RCMR_IN030000UK06.Message complex type.
- * 
- * &lt;p&gt;The following schema fragment specifies the expected content contained within this class.
- * 
- * &lt;pre&gt;
- * &amp;lt;complexType name="RCMR_IN030000UK06.Message"&amp;gt;
- *   &amp;lt;complexContent&amp;gt;
- *     &amp;lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&amp;gt;
- *       &amp;lt;sequence&amp;gt;
- *         &amp;lt;element name="id" type="{urn:hl7-org:v3}II"/&amp;gt;
- *         &amp;lt;element name="creationTime" type="{urn:hl7-org:v3}TS"/&amp;gt;
- *         &amp;lt;element name="versionCode" type="{urn:hl7-org:v3}CS"/&amp;gt;
- *         &amp;lt;element name="interactionId" type="{urn:hl7-org:v3}II"/&amp;gt;
- *         &amp;lt;element name="processingCode" type="{urn:hl7-org:v3}CS"/&amp;gt;
- *         &amp;lt;element name="processingModeCode" type="{urn:hl7-org:v3}CS"/&amp;gt;
- *         &amp;lt;element name="acceptAckCode" type="{urn:hl7-org:v3}CS"/&amp;gt;
- *         &amp;lt;element name="communicationFunctionRcv" type="{urn:hl7-org:v3}MCCI_MT010101UK12.CommunicationFunctionRcv" maxOccurs="unbounded"/&amp;gt;
- *         &amp;lt;element name="communicationFunctionSnd" type="{urn:hl7-org:v3}MCCI_MT010101UK12.CommunicationFunctionSnd"/&amp;gt;
- *         &amp;lt;element name="ControlActEvent" type="{urn:hl7-org:v3}RCMR_IN030000UK06.ControlActEvent"/&amp;gt;
- *       &amp;lt;/sequence&amp;gt;
- *       &amp;lt;attribute name="type" type="{urn:hl7-org:v3}Classes" default="Message" /&amp;gt;
- *       &amp;lt;attribute name="typeID"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}oid" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="realmCode"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}cs" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="nullFlavor" type="{urn:hl7-org:v3}cs" /&amp;gt;
- *     &amp;lt;/restriction&amp;gt;
- *   &amp;lt;/complexContent&amp;gt;
- * &amp;lt;/complexType&amp;gt;
- * &lt;/pre&gt;
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "RCMR_IN030000UK07.Message", propOrder = {
-    "id",
-    "creationTime",
-    "versionCode",
-    "interactionId",
-    "processingCode",
-    "processingModeCode",
-    "acceptAckCode",
-    "communicationFunctionRcv",
-    "communicationFunctionSnd",
-    "controlActEvent"
-})
-public class RCMRIN030000UK07Message implements RCMRIN030000UKMessage {
-
-    @XmlElement(required = true)
-    protected II id;
-    @XmlElement(required = true)
-    protected TS creationTime;
-    @XmlElement(required = true)
-    protected CS versionCode;
-    @XmlElement(required = true)
-    protected II interactionId;
-    @XmlElement(required = true)
-    protected CS processingCode;
-    @XmlElement(required = true)
-    protected CS processingModeCode;
-    @XmlElement(required = true)
-    protected CS acceptAckCode;
-    @XmlElement(required = true)
-    protected List<MCCIMT010101UK12CommunicationFunctionRcv> communicationFunctionRcv;
-    @XmlElement(required = true)
-    protected MCCIMT010101UK12CommunicationFunctionSnd communicationFunctionSnd;
-    @XmlElement(name = "ControlActEvent", required = true)
-    protected RCMRIN030000UK07ControlActEvent controlActEvent;
-    @XmlAttribute(name = "type")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String type;
-    @XmlAttribute(name = "typeID")
-    protected List<String> typeID;
-    @XmlAttribute(name = "realmCode")
-    protected List<String> realmCode;
-    @XmlAttribute(name = "nullFlavor")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String nullFlavor;
-
-    /**
-     * Gets the value of the id property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link II }
-     *     
-     */
-    public II getId() {
-        return id;
-    }
-
-    /**
-     * Sets the value of the id property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link II }
-     *     
-     */
-    public void setId(II value) {
-        this.id = value;
-    }
-
-    /**
-     * Gets the value of the creationTime property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link TS }
-     *     
-     */
-    public TS getCreationTime() {
-        return creationTime;
-    }
-
-    /**
-     * Sets the value of the creationTime property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link TS }
-     *     
-     */
-    public void setCreationTime(TS value) {
-        this.creationTime = value;
-    }
-
-    /**
-     * Gets the value of the versionCode property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link CS }
-     *     
-     */
-    public CS getVersionCode() {
-        return versionCode;
-    }
-
-    /**
-     * Sets the value of the versionCode property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link CS }
-     *     
-     */
-    public void setVersionCode(CS value) {
-        this.versionCode = value;
-    }
-
-    /**
-     * Gets the value of the interactionId property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link II }
-     *     
-     */
-    public II getInteractionId() {
-        return interactionId;
-    }
-
-    /**
-     * Sets the value of the interactionId property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link II }
-     *     
-     */
-    public void setInteractionId(II value) {
-        this.interactionId = value;
-    }
-
-    /**
-     * Gets the value of the processingCode property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link CS }
-     *     
-     */
-    public CS getProcessingCode() {
-        return processingCode;
-    }
-
-    /**
-     * Sets the value of the processingCode property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link CS }
-     *     
-     */
-    public void setProcessingCode(CS value) {
-        this.processingCode = value;
-    }
-
-    /**
-     * Gets the value of the processingModeCode property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link CS }
-     *     
-     */
-    public CS getProcessingModeCode() {
-        return processingModeCode;
-    }
-
-    /**
-     * Sets the value of the processingModeCode property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link CS }
-     *     
-     */
-    public void setProcessingModeCode(CS value) {
-        this.processingModeCode = value;
-    }
-
-    /**
-     * Gets the value of the acceptAckCode property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link CS }
-     *     
-     */
-    public CS getAcceptAckCode() {
-        return acceptAckCode;
-    }
-
-    /**
-     * Sets the value of the acceptAckCode property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link CS }
-     *     
-     */
-    public void setAcceptAckCode(CS value) {
-        this.acceptAckCode = value;
-    }
-
-    /**
-     * Gets the value of the communicationFunctionRcv property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the communicationFunctionRcv property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getCommunicationFunctionRcv().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link MCCIMT010101UK12CommunicationFunctionRcv }
-     * 
-     * 
-     */
-    public List<MCCIMT010101UK12CommunicationFunctionRcv> getCommunicationFunctionRcv() {
-        if (communicationFunctionRcv == null) {
-            communicationFunctionRcv = new ArrayList<>();
-        }
-        return this.communicationFunctionRcv;
-    }
-
-    /**
-     * Gets the value of the communicationFunctionSnd property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link MCCIMT010101UK12CommunicationFunctionSnd }
-     *     
-     */
-    public MCCIMT010101UK12CommunicationFunctionSnd getCommunicationFunctionSnd() {
-        return communicationFunctionSnd;
-    }
-
-    /**
-     * Sets the value of the communicationFunctionSnd property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link MCCIMT010101UK12CommunicationFunctionSnd }
-     *     
-     */
-    public void setCommunicationFunctionSnd(MCCIMT010101UK12CommunicationFunctionSnd value) {
-        this.communicationFunctionSnd = value;
-    }
-
-    /**
-     * Gets the value of the controlActEvent property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link RCMRIN030000UK06ControlActEvent }
-     *     
-     */
-    public RCMRIN030000UK07ControlActEvent getControlActEvent() {
-        return controlActEvent;
-    }
-
-    /**
-     * Sets the value of the controlActEvent property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link RCMRIN030000UK07ControlActEvent }
-     *     
-     */
-    public void setControlActEvent(RCMRIN030000UK07ControlActEvent value) {
-        this.controlActEvent = value;
-    }
-
-    /**
-     * Gets the value of the type property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getType() {
-        if (type == null) {
-            return "Message";
-        } else {
-            return type;
-        }
-    }
-
-    /**
-     * Sets the value of the type property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setType(String value) {
-        this.type = value;
-    }
-
-    /**
-     * Gets the value of the typeID property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the typeID property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getTypeID().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getTypeID() {
-        if (typeID == null) {
-            typeID = new ArrayList<>();
-        }
-        return this.typeID;
-    }
-
-    /**
-     * Gets the value of the realmCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the realmCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getRealmCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getRealmCode() {
-        if (realmCode == null) {
-            realmCode = new ArrayList<>();
-        }
-        return this.realmCode;
-    }
-
-    /**
-     * Gets the value of the nullFlavor property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNullFlavor() {
-        return nullFlavor;
-    }
-
-    /**
-     * Sets the value of the nullFlavor property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNullFlavor(String value) {
-        this.nullFlavor = value;
-    }
-
-}
