--- schema/src/main/java/org/hl7/v3/RCMRIN030000UK07Subject.java
@@ -1,262 +1,0 @@
-
-package org.hl7.v3;
-
-import java.util.ArrayList;
-import java.util.List;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlAttribute;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * &lt;p&gt;Java class for RCMR_IN030000UK06.Subject complex type.
- * 
- * &lt;p&gt;The following schema fragment specifies the expected content contained within this class.
- * 
- * &lt;pre&gt;
- * &amp;lt;complexType name="RCMR_IN030000UK06.Subject"&amp;gt;
- *   &amp;lt;complexContent&amp;gt;
- *     &amp;lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&amp;gt;
- *       &amp;lt;sequence&amp;gt;
- *         &amp;lt;element name="EhrExtract" type="{urn:hl7-org:v3}RCMR_MT030101UK04.EhrExtract"/&amp;gt;
- *       &amp;lt;/sequence&amp;gt;
- *       &amp;lt;attribute name="type" type="{urn:hl7-org:v3}Classes" default="ActRelationship" /&amp;gt;
- *       &amp;lt;attribute name="typeCode" type="{urn:hl7-org:v3}ActRelationshipType" default="SUBJ" /&amp;gt;
- *       &amp;lt;attribute name="contextConductionInd" type="{urn:hl7-org:v3}bl" default="false" /&amp;gt;
- *       &amp;lt;attribute name="typeID"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}oid" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="realmCode"&amp;gt;
- *         &amp;lt;simpleType&amp;gt;
- *           &amp;lt;list itemType="{urn:hl7-org:v3}cs" /&amp;gt;
- *         &amp;lt;/simpleType&amp;gt;
- *       &amp;lt;/attribute&amp;gt;
- *       &amp;lt;attribute name="nullFlavor" type="{urn:hl7-org:v3}cs" /&amp;gt;
- *     &amp;lt;/restriction&amp;gt;
- *   &amp;lt;/complexContent&amp;gt;
- * &amp;lt;/complexType&amp;gt;
- * &lt;/pre&gt;
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "RCMR_IN030000UK07.Subject", propOrder = {
-    "ehrExtract"
-})
-public class RCMRIN030000UK07Subject {
-
-    @XmlElement(name = "EhrExtract", required = true)
-    protected RCMRMT030101UKEhrExtract ehrExtract;
-    @XmlAttribute(name = "type")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String type;
-    @XmlAttribute(name = "typeCode")
-    protected List<String> typeCode;
-    @XmlAttribute(name = "contextConductionInd")
-    protected Boolean contextConductionInd;
-    @XmlAttribute(name = "typeID")
-    protected List<String> typeID;
-    @XmlAttribute(name = "realmCode")
-    protected List<String> realmCode;
-    @XmlAttribute(name = "nullFlavor")
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    protected String nullFlavor;
-
-    /**
-     * Gets the value of the ehrExtract property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link RCMRMT030101UKEhrExtract }
-     *     
-     */
-    public RCMRMT030101UKEhrExtract getEhrExtract() {
-        return ehrExtract;
-    }
-
-    /**
-     * Sets the value of the ehrExtract property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link RCMRMT030101UKEhrExtract }
-     *     
-     */
-    public void setEhrExtract(RCMRMT030101UKEhrExtract value) {
-        this.ehrExtract = value;
-    }
-
-    /**
-     * Gets the value of the type property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getType() {
-        if (type == null) {
-            return "ActRelationship";
-        } else {
-            return type;
-        }
-    }
-
-    /**
-     * Sets the value of the type property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setType(String value) {
-        this.type = value;
-    }
-
-    /**
-     * Gets the value of the typeCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the typeCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getTypeCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getTypeCode() {
-        if (typeCode == null) {
-            typeCode = new ArrayList<>();
-        }
-        return this.typeCode;
-    }
-
-    /**
-     * Gets the value of the contextConductionInd property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Boolean }
-     *     
-     */
-    public boolean isContextConductionInd() {
-        if (contextConductionInd == null) {
-            return false;
-        } else {
-            return contextConductionInd;
-        }
-    }
-
-    /**
-     * Sets the value of the contextConductionInd property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Boolean }
-     *     
-     */
-    public void setContextConductionInd(Boolean value) {
-        this.contextConductionInd = value;
-    }
-
-    /**
-     * Gets the value of the typeID property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the typeID property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getTypeID().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getTypeID() {
-        if (typeID == null) {
-            typeID = new ArrayList<>();
-        }
-        return this.typeID;
-    }
-
-    /**
-     * Gets the value of the realmCode property.
-     * 
-     * &lt;p&gt;
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a &lt;CODE&gt;set&lt;/CODE&gt; method for the realmCode property.
-     * 
-     * &lt;p&gt;
-     * For example, to add a new item, do as follows:
-     * &lt;pre&gt;
-     *    getRealmCode().add(newItem);
-     * &lt;/pre&gt;
-     * 
-     * 
-     * &lt;p&gt;
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     */
-    public List<String> getRealmCode() {
-        if (realmCode == null) {
-            realmCode = new ArrayList<>();
-        }
-        return this.realmCode;
-    }
-
-    /**
-     * Gets the value of the nullFlavor property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getNullFlavor() {
-        return nullFlavor;
-    }
-
-    /**
-     * Sets the value of the nullFlavor property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setNullFlavor(String value) {
-        this.nullFlavor = value;
-    }
-
-}
